requirements.txt
mcp>=1.0.0
pyyaml>=6.0
tiktoken>=0.5.0
transformers>=4.30.0
pydantic>=2.0.0
typing-extensions>=4.5.0

